+++ LookupAnything/release-notes.md
@@ -2,6 +2,7 @@
 
 # Release notes
 ## Upcoming release
+* Fixed unable to lookup an NPC from their profile menu.
 * Fixed NPC lookups showing gift tastes for flavored variants in 1.32.3.
 * Fixed tree lookups incorrectly stating that a fertilized tree won't grow in winter.
 * Fixed error when trying to lookup on the 'help wanted' screen.